requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
mysqlclient>=2.2,<3.0
django-environ>=0.11,<1.0
django-cors-headers>=4.3,<5.0
Pillow>=10.0
pyotp>=2.9
